Upgrade Time service to .NET 6
