+++ src/TimeService/GlobalUsings.cs
@@ -1,0 +1,7 @@
+global using Pitstop.TimeService;
+global using Pitstop.TimeService.Events;
+global using Microsoft.Extensions.DependencyInjection;
+global using Microsoft.Extensions.Hosting;
+global using Pitstop.Infrastructure.Messaging.Configuration;
+global using Pitstop.Infrastructure.Messaging;
+global using Serilog;
