+++ src/TimeService/Program.cs
@@ -1,52 +1,16 @@
-﻿using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Hosting;
-using Pitstop.Infrastructure.Messaging.Configuration;
-using Serilog;
-using System;
-using System.IO;
-using System.Threading;
-using System.Threading.Tasks;
-
-namespace Pitstop.TimeService
-{
-    class Program
+﻿IHost host = Host
+    .CreateDefaultBuilder(args)
+    .ConfigureServices((hostContext, services) =>
     {
-        public static async Task Main(string[] args)
-        {
-            var host = CreateHostBuilder(args).Build();
-            await host.RunAsync();
-        }
+        services.UseRabbitMQMessagePublisher(hostContext.Configuration);
 
-        private static IHostBuilder CreateHostBuilder(string[] args)
-        {
-            var hostBuilder = Host.CreateDefaultBuilder(args)
-                .ConfigureHostConfiguration(configHost =>
-                {
-                    configHost.SetBasePath(Directory.GetCurrentDirectory());
-                    configHost.AddJsonFile("hostsettings.json", optional: true);
-                    configHost.AddJsonFile($"appsettings.json", optional: false);
-                    configHost.AddEnvironmentVariables();
-                    configHost.AddEnvironmentVariables("DOTNET_");
-                    configHost.AddCommandLine(args);
-                })
-                .ConfigureAppConfiguration((hostContext, config) =>
-                {
-                    config.AddJsonFile($"appsettings.{hostContext.HostingEnvironment.EnvironmentName}.json", optional: false);
-                })
-                .ConfigureServices((hostContext, services) =>
-                {
-                    services.UseRabbitMQMessagePublisher(hostContext.Configuration);
+        services.AddHostedService<TimeWorker>();
+    })
+    .UseSerilog((hostContext, loggerConfiguration) =>
+    {
+        loggerConfiguration.ReadFrom.Configuration(hostContext.Configuration);
+    })
+    .UseConsoleLifetime()
+    .Build();
 
-                    services.AddHostedService<TimeManager>();
-                })
-                .UseSerilog((hostContext, loggerConfiguration) =>
-                {
-                    loggerConfiguration.ReadFrom.Configuration(hostContext.Configuration);
-                })
-                .UseConsoleLifetime();
-
-            return hostBuilder;
-        }        
-    }
-}
+await host.RunAsync();
